Move group params verifier
