+++ libs/membership/membership-common/src/main/java/net/corda/membership/lib/verifiers/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.membership.lib.verifiers;
+
+import org.osgi.annotation.bundle.Export;
