+++ libs/membership/membership-common/src/main/kotlin/net/corda/membership/lib/verifiers/GroupParametersUpdateVerifier.kt
@@ -1,4 +1,4 @@
-package net.corda.membership.lib.impl.verifiers
+package net.corda.membership.lib.verifiers
 
 import net.corda.membership.lib.GroupParametersNotaryUpdater.Companion.MPV_KEY
